[FX] Redundant sanity test is removed
